+++ src/rdb/lib/faccposix.cc
@@ -23,7 +23,7 @@ std::string posixBinaryFileAccessor<T>::fileName() {
 }
 
 template <class T>
-int posixBinaryFileAccessor<T>::write(const T* ptrData, const size_t size, const size_t position) {
+ssize_t posixBinaryFileAccessor<T>::write(const T* ptrData, const size_t size, const size_t position) {
   int fd;
   if (position == std::numeric_limits<size_t>::max()) {
     fd = ::open(fileNameStr.c_str(), O_APPEND | O_RDWR | O_CREAT | kOpenBaseFlags, 0644);
@@ -63,7 +63,7 @@ int posixBinaryFileAccessor<T>::write(const T* ptrData, const size_t size, const
 }
 
 template <class T>
-int posixBinaryFileAccessor<T>::read(T* ptrData, const size_t size, const size_t position) {
+ssize_t posixBinaryFileAccessor<T>::read(T* ptrData, const size_t size, const size_t position) {
   int fd = -1;
   fd = ::open(fileNameStr.c_str(), O_RDONLY | kOpenBaseFlags);
   assert(fd >= 0);
